# docs/swagger.yaml
basePath: /api/v12
definitions:
  github_com_QuentinRegnier_nubo-backend_internal_domain.CreatePostResponse:
    properties:
      post_id:
        type: integer
    type: object
  github_com_QuentinRegnier_nubo-backend_internal_domain.ErrorResponse:
    properties:
      error:
        example: Invalid JSON format
        type: string
    type: object
  github_com_QuentinRegnier_nubo-backend_internal_domain.LoginInput:
    type: object
  github_com_QuentinRegnier_nubo-backend_internal_domain.LoginResponse:
    properties:
      badges:
        items:
          type: string
        type: array
      ban_expires_at:
        type: string
      ban_reason:
        type: string
      banned:
        example: false
        type: boolean
      bio:
        example: Bio...
        type: string
      birthdate:
        type: string
      created_at:
        type: string
      desactivated:
        example: false
        type: boolean
      email:
        example: john@nubo.com
        type: string
      email_verified:
        example: true
        type: boolean
      expires_at:
        type: string
      first_name:
        example: John
        type: string
      grade:
        example: 0
        type: integer
      jwt:
        example: eyJhbGciOiJIUzI1Ni...
        type: string
      last_name:
        example: Doe
        type: string
      location:
        example: Paris
        type: string
      master_token:
        example: eyJhbGci...
        type: string
      message:
        example: Login successful
        type: string
      phone:
        example: "+33612345678"
        type: string
      phone_verified:
        example: true
        type: boolean
      school:
        example: "42"
        type: string
      sex:
        example: 1
        type: integer
      updated_at:
        type: string
      user_id:
        example: 42
        type: integer
      username:
        example: johndoe
        type: string
      work:
        example: Dev
        type: string
    type: object
...
  github_com_QuentinRegnier_nubo-backend_internal_domain.RenewJWTResponse:
    properties:
      message:
        type: string
      token:
        type: string
    type: object
  github_com_QuentinRegnier_nubo-backend_internal_domain.SignUpResponse:
    properties:
      expires_at:
        type: string
      jwt:
        example: eyJhbGciOiJIUzI1Ni...
        type: string
      master_token:
        example: eyJhbGciOiJIUzI1Ni...
        type: string
      message:
        example: User created successfully
        type: string
      profile_picture_id:
        type: integer
      user_id:
        example: 42
        type: integer
    type: object
info:
  contact: {}
  description: |-
    Documentation de l'API.
    Pour mettre à jour :
  title: Mon API Propre
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      - multipart/form-data
      description: "Authentifie un utilisateur via email/password et renvoie son profil
        complet + token.\n\n**Règles & Erreurs :**\n\n\U0001F534 **400 Bad Request
        :**\n* `The 'data' field containing the JSON is required` : Champ 'data' manquant.\n*
        `Invalid JSON format in 'data'` : Le JSON envoyé est mal formé.\n\n\U0001F7E0
        **401 Unauthorized :**\n* `Invalid email or password` : Identifiants incorrects
        ou utilisateur introuvable.\n\n⛔ **403 Forbidden :**\n* `Account deactivated`
        : Le compte a été désactivé.\n* `Account banned` : Le compte a été banni.\n\n⚫
        **500 Internal Server Error :**\n* `database error` : Erreur technique interne."
      parameters:
      - description: Données JSON (si multipart/form-data)
        in: formData
        name: data
        type: string
      - description: Données JSON (si application/json)
        in: body
        name: request
        schema:
          $ref: '#/definitions/github_com_QuentinRegnier_nubo-backend_internal_domain.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_QuentinRegnier_nubo-backend_internal_domain.LoginResponse'
        "400":
          description: Adresse IP invalide
          schema:
            $ref: '#/definitions/github_com_QuentinRegnier_nubo-backend_internal_domain.ErrorResponse'
        "401":
          description: Identifiants invalides
          schema:
            $ref: '#/definitions/github_com_QuentinRegnier_nubo-backend_internal_domain.ErrorResponse'
        "403":
          description: Compte bloqué/banni
          schema:
            $ref: '#/definitions/github_com_QuentinRegnier_nubo-backend_internal_domain.ErrorResponse'
        "500":
          description: Erreur serveur
          schema:
            $ref: '#/definitions/github_com_QuentinRegnier_nubo-backend_internal_domain.ErrorResponse'
      summary: Connecter un utilisateur
      tags:
      - users
  /post:
    post:
      consumes:
      - multipart/form-data
      description: "Crée un post_service avec du contenu texte, des hashtags, des mentions
        d'utilisateurs et entre 1 et 4 images.\nCette route nécessite une authentification
        par JWT et une signature HMAC valide.\n\n**Règles de validation & Erreurs
        :**\n\n\U0001F534 **400 Bad Request (Erreurs client) :**\n* `Field 'data'
        is required` : Le champ texte 'data' contenant le JSON est manquant.\n* `Invalid
        JSON: ...` : Le format JSON dans le champ 'data' est incorrect.\n* `Too many
        tags (max 10)` : Le nombre de hashtags ou d'utilisateurs tagués dépasse 10.\n*
        `Maximum 4 images allowed` : Vous avez tenté d'envoyer plus de 4 fichiers
        média.\n\n\U0001F7E0 **401 Unauthorized (Authentification) :**\n* `Utilisateur
        non identifié` : Le userID n'a pas pu être extrait du token JWT ou contexte
        manquant.\n* `Signature HMAC invalide` : (Géré par le middleware) La signature
        ne correspond pas au contenu.\n\n⚫ **500 Internal Server Error (Serveur) :**\n*
        `Failed to create post_service: ...` : Erreur lors de l'upload MinIO ou de l'insertion
        dans la file d'attente Redis (Queue)."
      parameters:
      - description: Bearer <votre_jwt>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Signature HMAC de la requête
        in: header
        name: X-Signature
        required: true
        type: string
      - description: Timestamp Unix de la requête
        in: header
        name: X-Timestamp
        required: true
        type: string
      - description: Images du post_service (1 à 4 fichiers)
        in: formData
        name: media
        type: file
      - description: Données JSON (domain.CreatePostInput)
        in: formData
        name: data
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_QuentinRegnier_nubo-backend_internal_domain.CreatePostResponse'
        "400":
          description: Données invalides ou trop de fichiers
          schema:
            $ref: '#/definitions/github_com_QuentinRegnier_nubo-backend_internal_domain.ErrorResponse'
        "401":
          description: Session expirée ou signature HMAC corrompue
          schema:
            $ref: '#/definitions/github_com_QuentinRegnier_nubo-backend_internal_domain.ErrorResponse'
        "500":
          description: Erreur interne de persistance
          schema:
            $ref: '#/definitions/github_com_QuentinRegnier_nubo-backend_internal_domain.ErrorResponse'
      summary: Créer une nouvelle publication
      tags:
      - posts
  /renew-jwt:
    post:
      consumes:
      - application/json
      description: "Génère un nouveau JWT pour l'utilisateur et effectue une rotation
        de sécurité des secrets (Ratchet).\nCette route est critique et nécessite
        une signature HMAC valide basée sur le secret actuel de la session.\n\n**Mécanisme
        :**\n1. Vérifie la signature HMAC du body avec les headers de sécurité.\n2.
        Identifie la session via l'ID utilisateur et le `X-Secret`.\n3. Calcule le
        prochain secret (N+1) et met à jour l'historique (Ratchet).\n4. Renvoie le
        nouveau JWT.\n\n**Règles & Erreurs :**\n\n\U0001F534 **400 Bad Request :**\n*
        `Erreur lecture body` : Impossible de lire le corps de la requête.\n* `Invalid
        JSON format` : Le JSON envoyé est mal formé.\n* `Headers de sécurité manquants`
        : Il manque `Authorization`, `X-Secret`, `X-Signature` ou `X-Timestamp`.\n\n\U0001F7E0
        **401 Unauthorized :**\n* `Signature HMAC invalide` : La signature ne correspond
        pas au contenu (tentative de falsification).\n* `Session invalide ou Secret
        incorrect` : Le secret fourni ne correspond à aucune session active pour cet
        utilisateur (ou désynchronisation Ratchet).\n\n⚫ **500 Internal Server Error
        :**\n* `Erreur génération token` : Échec de la création du JWT.\n* `Erreur
        rotation secrets` : Impossible de mettre à jour Redis (Ratchet bloqué)."
      parameters:
      - description: Bearer <Last_JWT>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Secret actuel de la session
        in: header
        name: X-Secret
        required: true
        type: string
      - description: Signature HMAC calculée
        in: header
        name: X-Signature
        required: true
        type: string
      - description: Timestamp de la requête
        in: header
        name: X-Timestamp
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_QuentinRegnier_nubo-backend_internal_domain.RenewJWTResponse'
        "400":
          description: Requête invalide
          schema:
            $ref: '#/definitions/github_com_QuentinRegnier_nubo-backend_internal_domain.ErrorResponse'
        "401":
          description: Authentification / Signature refusée
          schema:
            $ref: '#/definitions/github_com_QuentinRegnier_nubo-backend_internal_domain.ErrorResponse'
        "500":
          description: Erreur serveur critique
          schema:
            $ref: '#/definitions/github_com_QuentinRegnier_nubo-backend_internal_domain.ErrorResponse'
      summary: Renouveler le JWT (Ratchet Rotation)
      tags:
      - auth
  /signup:
    post:
      consumes:
      - multipart/form-data
      description: "Inscription complète avec upload d'avatar et données JSON.\n\n**Règles
        de validation & Erreurs :**\n\n\U0001F534 **400 Bad Request (Erreurs client)
        :**\n* `The 'data' field containing the JSON is required` : Tu as oublié d'envoyer
        le champ texte 'data'.\n* `Invalid JSON format in 'data': ...` : Ton JSON
        est mal écrit (virgule manquante, accolade, etc).\n* `Invalid date format.
        Expected format: ddmmaaaa` : La date de naissance n'est pas bonne.\n* `Gender
        must be 0, 1, 2, or null` : Tu as envoyé un entier invalide pour le sexe.\n*
        `Impossible to read image file` : Le fichier image est corrompu ou illisible.\n\n\U0001F7E0
        **409 Conflict (Doublons) :**\n* `This username is already taken` : Le pseudo
        est déjà en base.\n\n⚫ **500 Internal Server Error (Problèmes serveur) :**\n*
        `Internal error (image upload)` : MinIO est down ou mal configuré.\n* `Internal
        error (token generation)` : Problème avec la signature JWT.\n* `database error`
        : Postgres ou Mongo ne répondent pas."
      parameters:
      - description: Photo de profil (Image)
        in: formData
        name: profile_picture
        type: file
      - description: Données JSON (domain.SignUpInput)
        in: formData
        name: data
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_QuentinRegnier_nubo-backend_internal_domain.SignUpResponse'
        "400":
          description: Données invalides (Voir liste ci-dessus)
          schema:
            $ref: '#/definitions/github_com_QuentinRegnier_nubo-backend_internal_domain.ErrorResponse'
        "409":
          description: Conflit (Pseudo pris)
          schema:
            $ref: '#/definitions/github_com_QuentinRegnier_nubo-backend_internal_domain.ErrorResponse'
        "500":
          description: Erreur Serveur
          schema:
            $ref: '#/definitions/github_com_QuentinRegnier_nubo-backend_internal_domain.ErrorResponse'
      summary: Créer un compte utilisateur
      tags:
      - users
swagger: "2.0"